[QPay]API-getTradeToken:add trade pwd,API-newTrade:send push msg

+++ qplayApi/qplayApi/app/Services/QPayTradeService.php
@@ -7,11 +7,13 @@ namespace App\Services;
 
 use App\lib\CommonUtil;
 use App\lib\ResultCode;
+use App\lib\PushUtil;
 use App\Repositories\QPayTradeTokenRepository;
 use App\Repositories\QPayMemberPointRepository;
 use App\Repositories\QPayMemberRepository;
 use App\Repositories\QPayShopRepository;
 use App\Repositories\QPayTradeLogRepository;
+use DB;
 
 class QPayTradeService
 {
@@ -48,43 +50,57 @@ class QPayTradeService
     * Get Trade Token
     * @return trade token
     */
-    public function getTradeToken($uuid, $empNO, $price, $shopID)
+    public function getTradeToken($uuid, $empNO, $tradePWD, $price, $shopID)
     {
         $resultCode = ResultCode::_1_reponseSuccessful;
         $cerateTradeToken = false;
+
+        $user = CommonUtil::getUserInfoJustByUserEmpNo($empNO);
         $tradeTokenData = $this->qpayTradeTokenRepository->getTradeToken($uuid);
+        $qpayMemberData = $this->qpayMemberRepository->getQPayMemberInfo($user->row_id);
 
-        //No Trade Token exist
-        if (count($tradeTokenData) == 0) {
-            $cerateTradeToken = true;
-        } else {
-            //Check Trade Token Valid
-            if (intval($tradeTokenData[0]->trade_token_valid) > time()) {
-                //Valid
-                $newTradeToken = $tradeTokenData[0]->trade_token;
-                $newTradeTokenValid = $tradeTokenData[0]->trade_token_valid;
-            } else {
-                //Invalid
+        if (password_verify($tradePWD, $qpayMemberData["trade_password"])) {
+            //Trade Password Valid
+            //No Trade Token exist
+            if (count($tradeTokenData) == 0) {
                 $cerateTradeToken = true;
-                $this->qpayTradeTokenRepository->deleteTradeToken($uuid, $tradeTokenData[0]->trade_token);
+            } else {
+                //Check Trade Token Valid
+                if (intval($tradeTokenData[0]->trade_token_valid) > time()) {
+                    //Valid
+                    $newTradeToken = $tradeTokenData[0]->trade_token;
+                    $newTradeTokenValid = $tradeTokenData[0]->trade_token_valid;
+                } else {
+                    //Invalid
+                    $cerateTradeToken = true;
+                    $this->qpayTradeTokenRepository->deleteTradeToken($uuid, $tradeTokenData[0]->trade_token);
+                }
             }
+
+            if ($cerateTradeToken) {
+                $newTradeTokenValid = intval(time() + 30);
+                $newTradeToken = base64_encode(md5($uuid . $empNO . $tradePWD . $price . $shopID . $newTradeTokenValid));
+
+                $this->qpayTradeTokenRepository->newTradeToken($uuid, $newTradeToken, $newTradeTokenValid);
+            }
+
+            $result = [
+                "result_code" => $resultCode,
+                "message" => CommonUtil::getMessageContentByCode($resultCode),
+                "content" => [
+                    "trade_token" => $newTradeToken,
+                    "trade_token_valid" => $newTradeTokenValid
+                ]
+            ];
+        } else {
+            //Trade Password Invalid
+            $resultCode = ResultCode::_000929_tradePasswordIncorrect;
+
+            $result = [
+                "result_code" => $resultCode,
+                "message" => CommonUtil::getMessageContentByCode($resultCode)
+            ];
         }
-
-        if ($cerateTradeToken) {
-            $newTradeTokenValid = intval(time() + 30);
-            $newTradeToken = base64_encode(md5($uuid . $empNO . $price . $shopID . $newTradeTokenValid));
-
-            $this->qpayTradeTokenRepository->newTradeToken($uuid, $newTradeToken, $newTradeTokenValid);
-        }
-
-        $result = [
-            "result_code" => $resultCode,
-            "message" => CommonUtil::getMessageContentByCode($resultCode),
-            "content" => [
-                "trade_token" => $newTradeToken,
-                "trade_token_valid" => $newTradeTokenValid
-            ]
-        ];
 
         return $result;
     }
@@ -93,7 +109,7 @@ class QPayTradeService
     * New Trade
     * @return trade result
     */
-    public function newTrade($uuid, $tradePWD, $tradeToken, $empNO, $price, $shopID)
+    public function newTrade($uuid, $tradePWD, $tradeToken, $empNO, $price, $shopID, $lang)
     {
         $newTradeID = "";
         $checkSuccess = true;
@@ -169,73 +185,32 @@ class QPayTradeService
 
         //Step 5. New Trade
         if ($checkSuccess) {
-            $allPointData = $this->qpayMemberPointRepository->getPointData($user->row_id);
+            DB::beginTransaction();
 
-            if ($allPointData[0]->stored_now > $price) {
-                //Not a Multiple Point Trade
-                $multiplePay = "N";
-                $multipleRowID = "";
-                $multiplePoint = "";
-                $tradeSuccess = "Y";
-                $newStoredNow = $allPointData[0]->stored_now - $price;
-                $newStoredUsed = $allPointData[0]->stored_used + $price;
+            try {
+                $allPointData = $this->qpayMemberPointRepository->getPointData($user->row_id);
 
-                $updatePointDataResult = $this->qpayMemberPointRepository->updatePointData($allPointData[0]->row_id, $newStoredNow,  $newStoredUsed);
+                if ($allPointData[0]->stored_now > $price) {
+                    //Not a Multiple Point Trade
+                    $multiplePay = "N";
+                    $multipleRowID = "";
+                    $multiplePoint = "";
+                    $tradeSuccess = "Y";
+                    $newStoredNow = $allPointData[0]->stored_now - $price;
+                    $newStoredUsed = $allPointData[0]->stored_used + $price;
 
+                    $updatePointDataResult = $this->qpayMemberPointRepository->updatePointData($allPointData[0]->row_id, $newStoredNow,  $newStoredUsed);
 
-                if ($updatePointDataResult != 1) {
-                    $tradeSuccess = "N";
-                }
-
-                $tradeID = $this->qpayTradeLogRepository->newTradeRecord(
-                    $allPointData[0]->member_row_id,
-                    $allPointData[0]->row_id,
-                    $allPointData[0]->stored_now,
-                    $allPointData[0]->stored_used,
-                    $newStoredNow,
-                    $newStoredUsed,
-                    $shopID,
-                    $multiplePay,
-                    $multipleRowID,
-                    $multiplePoint,
-                    $price,
-                    $tradeSuccess,
-                    $resultCode
-                );
-
-                $newTradeID = $tradeID;
-            } else {
-                //Multiple Point Trade
-                $multiplePay = "Y";
-                $multipleRowID = "";
-                $multiplePoint = "";
-                $tradeSuccess = "Y";
-                $dataProcessNumber = 1;
-                $tradePriceLeft = $price;
-
-                foreach ($allPointData as $pointData) {
-                    if ($dataProcessNumber == 1) {
-                        $newStoredNow = 0;
-                        $newStoredUsed = $pointData["stored_used"] + $pointData["stored_now"];
-                        $multiplePoint = $pointData["stored_now"];
-                        $tradePriceLeft = $price - $pointData["stored_now"];
-                    } else if ($dataProcessNumber == 2) {
-                        $newStoredNow = $pointData["stored_now"] - $tradePriceLeft;
-                        $newStoredUsed = $pointData["stored_used"] + $tradePriceLeft;
-                        $multiplePoint = $tradePriceLeft;
-                    }
-
-                    $updatePointDataResult = $this->qpayMemberPointRepository->updatePointData($pointData["row_id"], $newStoredNow,  $newStoredUsed);
 
                     if ($updatePointDataResult != 1) {
                         $tradeSuccess = "N";
                     }
 
-                    $latestPointTradeLogID = $this->qpayTradeLogRepository->newTradeRecord(
-                        $pointData["member_row_id"],
-                        $pointData["row_id"],
-                        $pointData["stored_now"],
-                        $pointData["stored_used"],
+                    $tradeID = $this->qpayTradeLogRepository->newTradeRecord(
+                        $allPointData[0]->member_row_id,
+                        $allPointData[0]->row_id,
+                        $allPointData[0]->stored_now,
+                        $allPointData[0]->stored_used,
                         $newStoredNow,
                         $newStoredUsed,
                         $shopID,
@@ -247,19 +222,87 @@ class QPayTradeService
                         $resultCode
                     );
 
-                    if ($dataProcessNumber == 1) {
-                        $multipleRowID = $latestPointTradeLogID;
-                        $dataProcessNumber++;
+                    $newTradeID = $tradeID;
+                } else {
+                    //Multiple Point Trade
+                    $multiplePay = "Y";
+                    $multipleRowID = "";
+                    $multiplePoint = "";
+                    $tradeSuccess = "Y";
+                    $dataProcessNumber = 1;
+                    $tradePriceLeft = $price;
+
+                    foreach ($allPointData as $pointData) {
+                        if ($dataProcessNumber == 1) {
+                            $newStoredNow = 0;
+                            $newStoredUsed = $pointData["stored_used"] + $pointData["stored_now"];
+                            $multiplePoint = $pointData["stored_now"];
+                            $tradePriceLeft = $price - $pointData["stored_now"];
+                        } else if ($dataProcessNumber == 2) {
+                            $newStoredNow = $pointData["stored_now"] - $tradePriceLeft;
+                            $newStoredUsed = $pointData["stored_used"] + $tradePriceLeft;
+                            $multiplePoint = $tradePriceLeft;
+                        }
+
+                        $updatePointDataResult = $this->qpayMemberPointRepository->updatePointData($pointData["row_id"], $newStoredNow,  $newStoredUsed);
+
+                        if ($updatePointDataResult != 1) {
+                            $tradeSuccess = "N";
+                        }
+
+                        $latestPointTradeLogID = $this->qpayTradeLogRepository->newTradeRecord(
+                            $pointData["member_row_id"],
+                            $pointData["row_id"],
+                            $pointData["stored_now"],
+                            $pointData["stored_used"],
+                            $newStoredNow,
+                            $newStoredUsed,
+                            $shopID,
+                            $multiplePay,
+                            $multipleRowID,
+                            $multiplePoint,
+                            $price,
+                            $tradeSuccess,
+                            $resultCode
+                        );
+
+                        if ($dataProcessNumber == 1) {
+                            $multipleRowID = $latestPointTradeLogID;
+                            $dataProcessNumber++;
+                        }
                     }
+
+                    $newTradeID = $multipleRowID;
                 }
 
-                $newTradeID = $multipleRowID;
-            }
+                if ($tradeSuccess) {
+                    $resultCode = ResultCode::_1_reponseSuccessful;
 
-            if ($tradeSuccess) {
-                $resultCode = ResultCode::_1_reponseSuccessful;
-            }
+                    //Send Push Message
+                    $shopData = $this->qpayShopRepository->getShopInfoByShopID($shopID);
+                    date_default_timezone_set("Asia/Taipei");
+                    $successTradeID = "T".str_pad($newTradeID, 6, "0", STR_PAD_LEFT);
 
+                    $queryParam =  array(
+                        'lang' => $lang
+                    );
+                    //$form = $shopData[0]->user_domain . "\\" . $shopData[0]->login_id;
+                    $form = PushUtil::getPushUserByEmpNo($empNO);
+                    $to = array(
+                        PushUtil::getPushUserByEmpNo($empNO),
+                        $shopData[0]->user_domain . "\\" . $shopData[0]->login_id
+                    );
+                    $title = trans("messages.MSG_QPAY_1");
+                    $text = str_replace("%0", date("m/d h:i"), trans("messages.MSG_QPAY_2"));
+                    $text = str_replace("%1", $successTradeID, $text);
+                    $extra = [];
+
+                    PushUtil::sendPushMessage($form, $to, $title, $text, $extra, $queryParam);
+                }
+            } catch (\Exception $e) {
+                DB::rollBack();
+                throw $e;
+            }
         } else {
             //Step 6. Create Trade Fail Log
             $tradeSuccess = "N";
